model: move MacroInterpreter to x86 model
